# Domashna3/backend/target/generated-sources/src/main/java/com/tunein/tuneinbackend/api/MoodApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.tunein.tuneinbackend.api;

import com.tunein.tuneinbackend.models.MoodRequest;
import com.tunein.tuneinbackend.models.Playlists;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-12-17T23:24:00.447004300+01:00[Europe/Berlin]")
@Api(value = "mood", description = "the mood API")
public interface MoodApi {

    @ApiOperation(value = "Get playlist based on mood", nickname = "getMoodPlaylist", notes = "", response = Playlists.class, tags={ "mood", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Playlists.class),
        @ApiResponse(code = 405, message = "invalid input"),
        @ApiResponse(code = 404, message = "not found") })
    @RequestMapping(value = "/mood",
        produces = { "application/json; charset=utf-8" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Playlists> getMoodPlaylist(@ApiParam(value = "" ,required=true )  @Valid @RequestBody MoodRequest body
);

}